refactor: history uses hooks
